requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autopep8==2.3.1
certifi==2024.12.14
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.4
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
exceptiongroup==1.2.2
Faker==33.1.0
gprof2dot==2024.6.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openai==1.58.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
toposort==1.10
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
